autonomous_research_agent/requirements.txt
click==8.1.3
python-dotenv==1.0.0
tqdm==4.65.0
jinja2==3.1.2
markdown==3.4.3
weasyprint==60.0
arxiv==1.4.8
requests==2.31.0
spacy==3.5.3
transformers==4.30.2
pipeline==0.1.0
numpy>=1.19.0
pandas>=1.3.0

# Need to run: python -m spacy download en_core_web_lg